【I4G54K】【AppStore】应用上传对应用包做完整性校验 --- 对于容器应用应该先校验再修改包文件，增加日志

+++ src/main/java/org/edgegallery/appstore/interfaces/app/facade/AppServiceFacade.java
@@ -308,6 +308,10 @@ public class AppServiceFacade {
             }
 
             if (!isImgZipExist) {
+                if (!appUtil.checkPackageValid(fileParent)) {
+                    throw new AppException("the app package is illegal and may have been tampered!",
+                        ResponseConst.RET_PACKAGE_ILLEGAL);
+                }
                 FileUtils.forceDelete(new File(fileAddress));
                 appService.updateAppPackageWithRepoInfo(fileParent);
                 appService.updateImgInRepo(imgDecsList);
